requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
bcrypt==4.1.1
Pillow>=10.0.0
requests==2.31.0
